requirements.txt
Django==5.1.4
psycopg2==2.9.10
django-redis==5.4.0
redis==5.2.1
djangorestframework==3.15.2